Finish chapt02-1
