--- learning_NN.py
@@ -1,26 +1,0 @@
-import chainer
-import chainer.function as func #F
-import chainer.links as links #L
-from   chainer import training, datasets, iterators, optimizers
-from   chainer.training import extensions
-import numpy as np
-
-class NMINT_Conv_NN(chainer.Chain):
-
-
-    def __init__(self):
-        super(NMINT_Conv_NN, self).__init__()
-        with self.init_scope():
-            self.conv1 = links.Convolution2D(1, 8, ksize = 3) #フィルタサイズ 3, 出力数8
-            self.linear1 = links.Linear(1352, 10) # 出力数10
-
-
-    def __call__(self, x):
-        #畳み込みニューラルネットワークによる画像認証
-        h1 = self.conv1(x)         # 畳み込み層
-        h2 = func.relu(h1)         # 活性化関数
-        h3 = max_pooling_2d(h2, 2) # プーリング層
-        h4 = self.linear1(h3)      # 全結合層
-
-        # 損失か結果を返す
-        return func.softmax_cross_entropy(h4, t) if train else func.softmax(h4)
